Synchronous call support
Introduced synchronous call support by adding sync() method to Request class.
When called during the flow API, a flag will be set to indicate that sync mode is to be use.
This flag is check after the request submition and, if flagged, then future will make use of get method to wait for request completion.