Add a code of conduct
As per ethics.org, a code of conduct has value as both an internal guideline and an external statement of community values and commitments and is a central guide and reference for contributors.

As there is no public email on Github, I've chosen the only public way for private communication with the main owner of this repo.