Add mock-server and base class
The mock server package is a standard for maven (and other) based projects to mock http requests at the connection level. You can run it as a standalone application or inject it into your runtime.

httpbin is very nice, but there are issues with having that in the test cases:
1. you need an internet connection to check the validity of new and current code
2. requests fail making tests undeterministic
3. the service is free and may stop existing at any point
4. adds significant time to the testing step
5. you don't win anything over injecting/mocking at the connection level