:100: mock all the things
This commit mocks *all* the requests instead of using httpbin and other online endpoints. Additionally:

- where possible, it uses the destructuring of components
- where applicable, it fixes the test by matching the testname (test named get uses method named get)
- where redundant, it removes the test as it already exists,
- where applicable, it moves the test into the right file

The result should be a speed up of over 90% and no failures due to connection issues or remote server issues.

Only one test is turned off, because it is actually faulty. Cancel requests does not seem to work consistently.