Testing: mock all the things (#434)
* Add mock-server and base class

The mock server package is a standard for maven (and other) based projects to mock http requests at the connection level. You can run it as a standalone application or inject it into your runtime.

httpbin is very nice, but there are issues with having that in the test cases:
1. you need an internet connection to check the validity of new and current code
2. requests fail making tests undeterministic
3. the service is free and may stop existing at any point
4. adds significant time to the testing step
5. you don't win anything over injecting/mocking at the connection level

* Add lower-case variant and TODO

Currently the request and response headers are cases sensitive, whereas per the RFCs they should be normalized by terms of making them case insensitive. We don't need to transform the headers, just make sure that access is case indifferent.

* :100: mock all the things

This commit mocks *all* the requests instead of using httpbin and other online endpoints. Additionally:

- where possible, it uses the destructuring of components
- where applicable, it fixes the test by matching the testname (test named get uses method named get)
- where redundant, it removes the test as it already exists,
- where applicable, it moves the test into the right file

The result should be a speed up of over 90% and no failures due to connection issues or remote server issues.

Only one test is turned off, because it is actually faulty. Cancel requests does not seem to work consistently.