Add mock server to all the submodules, remove jackson from coroutines and extract MockHelper
MockHelper does not need to be used as a TestCase and can therefore easily be injected in all types of test cases and runners. Currently could not find a way to be able to depend on it in other modules, so copying the file right now, but as they are identical, they could be removed once that is in place.