Mock fuel-android tests
Mocks fuel android tests without changes test outcome or what is under test.

I had to go JsonReader here, as kotlin wouldn't compile. It kept saying that the inferred type was of Any?, Iterator<Any?>, Sequene<Any?> or Map<Any?, String>. This always works, as it tokenizes the json and explicitely walks it.