Mock all the projects (#444)
* Add mock server to all the submodules, remove jackson from coroutines and extract MockHelper

MockHelper does not need to be used as a TestCase and can therefore easily be injected in all types of test cases and runners. Currently could not find a way to be able to depend on it in other modules, so copying the file right now, but as they are identical, they could be removed once that is in place.

* Mock fuel-android tests

Mocks fuel android tests without changes test outcome or what is under test.

I had to go JsonReader here, as kotlin wouldn't compile. It kept saying that the inferred type was of Any?, Iterator<Any?>, Sequene<Any?> or Map<Any?, String>. This always works, as it tokenizes the json and explicitely walks it.

* Mock fuel-coroutines test

Mocks fuel coroutines tests without changes to test outcome or what is under test. Removes the jackson dependency as it is completely unnecessary to test deserialization.

* Mock fuel-forge tests

Mocks fuel forge tests without changes to test outcome or what is under test. Removes the dependency on github issues and instead implements that also as a mocked call.

* Mock fuel-gson tests

Mocks fuel gson tests without changes to test outcome or what is under test. Removes the dependency on github issues and instead implements that also as a mocked call.

* Mock fuel-jackson tests

Mocks fuel jackson tests without changes to test outcome or what is under test. Removes the dependency on github issues and instead implements that also as a mocked call.

* Mock fuel-livedata tests

Mocks fuel livedata tests without changes test outcome or what is under test.

* Mock fuel-moshi tests

Mocks fuel moshi tests without changes to test outcome or what is under test. Removes the dependency on github issues and instead implements that also as a mocked call.

* Remove whitespace

Classes should not have 2 consecutive newlines and identation should be consistend throughout.

* Mock fuel-rxjava tests

Mocks fuel rxjava tests without changes test outcome or what is under test.

* Fix automatically renamed function