Make bodies be consumable at most once
This solves multiple issues with input streams being read twice, when logging is turned on. Before, toByteArray would load into memory and that was that. Now it reads directly from an inputstream. However, most streams are only consumable once, thus this solves that.

When the stream is read into memory we might as well keep it around.