Extract request options to executor, use composition for conditionally available methods, and actually make requests cancelable
This is a large commit as it overhauls part of the execution of the apis, mostly only affecting internals. I've tried to match current API as much as possible to reduce the churn because of this commit.

- Extracts request client options to the executor, instead of having them live on request. They only make sense conditionally and are not actually part of the request themselves. We keep growing the list of attributes, and this solves that.
- Composition for Download and Upload by wrapping request and overriding "this" (by using the "request" property). This allows us to conditionally make functions available and not rely on an enumeration. This also opens up a request that uploads AND downloads.
- Request cancelability was actually not implemented correctly. Some requests are currently still not cancellable! This change actually only exposes .cancel if the request is.

Fixes #399
Fixes #416
Fixes #344
Fixes #193
Should allow to implement #463 with an interceptor
Should allow to implement #464 with an interceptor
Should allow to implement #146 with a wrapped stream / interceptor