Refactor requests for request features
A lot of the parameters currently tracked by Request are actually options that are only used for the exection (and used by Client such as HttpClient). This change extracts these properties into executionOptions.

Additionally it implements downloads and uploads as features, instead of branches. This means that requests can do both instead of one of either. Fixes #510.

Finally it also makes sure that only requests that can be cancelled are exposing .cancel. Fixes #508. This also makes sure that #509 can be implemented.