Rich error stack traces and data handling
Expect the fuel errors to be rewrapped when it's recatched. This allows us to see all the bubbles (rethrows). The original error that yielded the fuel error is correctly stored (as the cause) and listed when the error is printed or stringified. For each consecutive throw, it only adds the new stack trace elements, which is great with asynchronous calls, as threads have their own stack.

Additionally adds explicit "Bubble" exception, and uses property getters instead of explicit passing. We know what response is.