Make FuelManager more usable and DRY-up
This change moves the Fuel class methods to an interface and delegates it to the default FuelManager. This means that all convenience methods are now also available on the FuelManager.

This is much easier to use, and makes it easier to have multiple managers, as RequestFactory or RequestFactory.Convenience can be used across the board.