Composable Request Features (#512)
* Make decoding and encoding reusable

There is no good reason for this to be inside HttpClient as it applies to any Input/OutputStream. This also reduces the complexity of HttpClient.

* Add kdoc for extensions

* Correctly serialize types for gson deserialization

Allows anything that Gson can serialize to be serialized directly as JSON body, and fixes deserialization of generic types.

Fixes #505

* Extens function extensions to support all results, make cancellable and async

Fixes #507

* Make monoResult expect something async

It was already implemented as such, but now it explicitly says it wants something async by using CancellableRequest.

Fixes #511

* Expect something async

Fixes #511

* Refactor requests for request features

A lot of the parameters currently tracked by Request are actually options that are only used for the exection (and used by Client such as HttpClient). This change extracts these properties into executionOptions.

Additionally it implements downloads and uploads as features, instead of branches. This means that requests can do both instead of one of either. Fixes #510.

Finally it also makes sure that only requests that can be cancelled are exposing .cancel. Fixes #508. This also makes sure that #509 can be implemented.

* Use extension function (for now)

* Refactor tests to actually test async

Fixes #509

* Allow wrapping in FuelError

Many times we try to cast exceptions into FuelErrors and then wrap them if it fails. This moves that logic to the FuelError class.

* Fix naming of classes

- The TaskRequest is actually a Task for a Request, so should be called RequestTask.
- The AsyncTaskRequest isn't actually Async, but attached callbacks to a RequestTask (thus is now called RequestTaskCallbacks).
- SuspendingRequest was decent, but it's not actually suspending until await is called, therefore it's now SuspendableRequest.

* Use extension for authentication

This allows users to easily add their own authentication methods that are only available when calling .authentication()

* Add type aliases for default responses

* Add toString overrides for wrapping tracking

* Implement all variants of coroutines

The issue here is that the resulting types were conflicting partially with core. This change actually allows you to get all the types, with the ease for non-wrapped results (T or Result<T, FuelError>).

We SHOULD also implement this for core.

* Allow more types of handlers to remove boilerplate

This allows for handlers that don't want request and response, remove the need to write _, _ in callbacks, all the time.

* Add and refactor tests

* Add cancel test

* Use MockHttpTestCase

Remove the need for testMode which actually didn't do anything.

* Remove unused test configuration

This is no longer being used anywhere.

* Make constructor private

There is nothing to instantiate here, this makes sure no one can.

* Rich error stack traces and data handling

Expect the fuel errors to be rewrapped when it's recatched. This allows us to see all the bubbles (rethrows). The original error that yielded the fuel error is correctly stored (as the cause) and listed when the error is printed or stringified. For each consecutive throw, it only adds the new stack trace elements, which is great with asynchronous calls, as threads have their own stack.

Additionally adds explicit "Bubble" exception, and uses property getters instead of explicit passing. We know what response is.

* Add documentation for handlers

* Always catch and wrap errors to track where they went wrong

* Automatically extract data out of status code not in range error

* Add method to printing

* Allow passing in url for tracking where the response should have been from

* Actually handle cancellations correctly

We must capture interruptions ourselve. That is how thread interruptions work. This allows us to interrupt during output, input and inbetween.

* :nail_care: lint

* Move outside of string builder for clearer intent

The "title" should be one interpolated string.

* :book: kdoc

* Fix InputStream failures

When inputStream creation fails, per the documentation, it may throw an IOException. This ensure we use the errorStream in case of such an error, and that, upon error, the streams are closed.

* Let the stream live if there is no issue

Finally closes the stream upon leaving the block, as it should, but we don't want that.

* Add kdoc and default to IO

IO dispatchers should be used for Blockin I/O, such as network related requests

* Add all the coroutine tests

* Add normalised tests for all response types, sync and async

* Make FuelManager more usable and DRY-up

This change moves the Fuel class methods to an interface and delegates it to the default FuelManager. This means that all convenience methods are now also available on the FuelManager.

This is much easier to use, and makes it easier to have multiple managers, as RequestFactory or RequestFactory.Convenience can be used across the board.

* Remove BaseTestCase

It's not good for anything anymore.

* Cleanup indentation

* Fix default parameters for request base abstract functions

* Fix coroutine package

* Remove JVM support

* Extract :scissors: UploadBody from UploadRequest

* Extract :scissors: to DefaultBody

* :nail_care: lint

* Documentation for 2.0.0 (#534)

* :scissors: cut into separate files

* Fix parameter encoding for GET requests

* Hello new :sparkles: documentation

* Some tests for documentation integrity

* Reset those FuelManagers

The issue is that if the FuelManager.instance is changed, tests are no longer running in isolation. We must ALWAYS reset the FuelManager after it has been modified, as this is a Singleton across tests.

* Update versions based on Constants.kt

* Replace prinln with trace

This allows us to add debugging information that is not executed when it's tracing is not active.

* Remove explicit JVM support

* Fix test-method

* Unify Request/Response toString (#535)

Make body parens consistent (don't include parens when body isn't empty)
Remove un-needed quotes