As per the RFCs, 307 and 308 should re-use the body (#548)
* As per the RFCs, 307 and 308 should re-use the body

This change makes bodies repeatable, which is turned on for in-memory values, and turned off for streams by default.

Repeatable bodies are _not_ consumed upon writing and can therefore be sent again on certain redirections.

* :nail_care: lint

* Fix redirection with authorization in base headers

This fixes an issue with redirection to a new host with base headers set for the current host.

* Actually fix the issue

The problem was that if you remove the header from the headers that are being injected, the header simply isn't injected. But it's a base header! So it's already there.

Now it uses a side-effect (also) to actually remove it.