Move encoding to ParameterEncoder
This enables encoding parameters in urls correctly for ALL request types, including array parameters, null parameters, empty parameters.

Also allows for #546 without any extra configuration. Try to put the parameters in the body, but if there's already a body, put it in the URL. Still allows for "pre-encoded" urls.

Fixes #546