Overhaul upload requests, Blob and DataPart
Based on RFC7578: https://tools.ietf.org/html/rfc7578

This is a breaking change, but mostly deprecates current functions as error.

Removes all the extranous methods and bookkeeping from UploadRequest and UploadBody, and delegates those to the DataPart. This also allows for InlineDataPart that yield a file on the server and removes automatic naming for dataparts (_1, _2), because the correct way to put multiple files under the same name is to give them... the same name!

Fixes #344
Fixes #390
Fixes #399
Fixes #479
Fixes #513