Parameter encoding for all requests (#553)
* Extract body representation into file and whitelist instead of blacklist

This whitelists text types as well as known application text representations to display when logging the body. Anything else, including custom and vendor media types that are not explicitly text are displayed as (x bytes of y).

* :nail_care: lint it

* Repeatable Bodies should be printable

* Set default content type if not yet set.

* Show encoded parameters in body

* Use static interceptors for logging

* Use typealiases

* Move encoding to ParameterEncoder

This enables encoding parameters in urls correctly for ALL request types, including array parameters, null parameters, empty parameters.

Also allows for #546 without any extra configuration. Try to put the parameters in the body, but if there's already a body, put it in the URL. Still allows for "pre-encoded" urls.

Fixes #546

* Update documentation to reflect changes

* Remove redundant tests and fix encoding header injection

* :police_car: nits :dash:

* :nail_care: remove empty line