Fix Jackson extension functions and their tests (#777)
* Consistent return type in extensions (#775)

Make all Jackson shorthand functions for asynchronous requests return
an object of `CancellableRequest` type.

* Fix tests for async Jackson extensions (#776)

Assertions in the tests were failing silently due to the fact that
the callback was executed asynchronously. Many tests eventually became
out of date, and because of the async assertions this remained unnoticed.

This change awaits on the result of each asynchronous call to ensure all
assertions in the asynchronous callbacks are executed. Should an error
occur, it will bubble up and hence fail the test now.

This also required quite some changes in the setup procedure and assertions
to bring tests back into a good shape.