Fixing forceMethod extension
When using HTTPS, the class hierarchy for the connection class is more complex. It is necessary to go back up in the parent classes in order to identify the class containing the 'method' field to update.

Hence the new recursive method allowing to go up the hierarchy of the current connection class.